1. Start discovery process         2. Change Outputdevice struct         3. Start cast process and listener, modified provider map to providerManager map         4. CastController, play_info_holder and play_info         5. Others         6. New interface_sdk_js         7. Ready to merge, add callbacks
        modified:
frameworks/js/napi/session/include/napi_avcast_controller.h
        modified:
frameworks/js/napi/session/include/napi_avcast_controller_callback.h
        modified:
frameworks/js/napi/session/src/napi_avcast_controller_callback.cpp
        modified:
frameworks/native/session/include/avcast_controller_callback_client.h
        modified:
frameworks/native/session/src/avcast_controller_callback_client.cpp
        modified:
interfaces/inner_api/native/session/include/avsession_info.h
        modified:
services/session/ipc/proxy/avcast_controller_callback_proxy.cpp
        modified:
services/session/ipc/proxy/avcast_controller_callback_proxy.h
        modified:
services/session/ipc/stub/avcast_controller_callback_stub.cpp
        modified:   services/session/server/avcast_controller_item.cpp
        modified:   services/session/server/avcast_controller_item.h

+++ frameworks/js/napi/session/include/napi_avcast_controller.h
@@ -83,7 +83,7 @@ private:
     std::string surfaceId_;
     int32_t volume_;
     int32_t loopMode_;
-    int32_t playSpeed_;
+    double playSpeed_;
     int32_t currentTime_;
     std::shared_ptr<AVCastController> castController_;
     std::shared_ptr<NapiAVCastControllerCallback> callback_;
